Add tests for social views and helper functions
