+++ tests/test_data/case/social_case_unknown_eligibility_status.json
@@ -1,0 +1,21 @@
+{
+  "state": "INACTIONABLE",
+  "id": "8849c299-5014-4637-bd2b-fc866aeccdf5",
+  "actionPlanId": "e6ce828c-ed94-457e-85a6-bcef784f4160",
+  "collectionInstrumentId": "724f7d1c-4343-408b-be06-725422a2e223",
+  "iac": "012345678910",
+  "caseRef": "1000000000000502",
+  "createdBy": "System",
+  "sampleUnitId": "519bb700-1bd9-432d-9db7-d34ea1727415",
+  "sampleUnitType": "H",
+  "createdDateTime": "2017-10-16T12:19:22.035+0000",
+  "caseGroup": {
+    "collectionExerciseId": "6553d121-df61-4b3a-8f43-e0726666b8cc",
+    "id": "26d65fa5-c8c3-4cad-bab0-44a52f81ca42",
+    "sampleUnitRef": "LMS0001",
+    "caseGroupStatus": "UNKNOWNELIGIBILITY",
+    "sampleUnitType": "H"
+  },
+  "responses": [],
+  "caseEvents": null
+}
